Simplify plugins significantly
Instead of having a complicated hierarchy with templated interfaces per element
type and such, there is now a single plugin type implementing an interface with
a single method. This also means that one plugin can contain elements of
different types, e.g. a DataSource and SampleDecoders. (I don't expect that
feature to be particularly useful; it's simply a side effect of the new
architecture.)

The only thing that a plugin can do is return a list of QMetaObjects
representing the classes that the plugin contains. It is up to the PluginManager
to actually instantiate any of these classes.