Use QtGstreamer instead of plain GStreamer
It's much nicer to use a C++ interface and reduces the potential for
memory-management errors. This does introduce an external dependency, but that's
not too big a deal. It also involves a downgrade from GStreamer 1.0 to 0.10
since QtGStreamer hasn't been released for the 1.0 series yet; again, I don't
think that matters.