Update EegDecoder for multiple chips
Also changed the format of EegSamples: removed fields we don't care about and
increased the width of the status fields appropriately. We allocate a lot of
these, so this should save us a little bit of memor. The width of the lead-off
status field will need to be manually adjusted if the software is ever used with
a device with nChannels != 16.