Implement function to wrap data in a QGst::Buffer
This is a bit trickier than it sounds. The function adds the data to a cache
which, because QByteArrays are implicitly shared, means that no copying is
done and the data is prevented from being freed while the buffer holds a
reference to it. It is removed from the cache when the buffer is destroyed.