Implement bufferFromBytes variant that copies data
I keep getting crashes because the internal data cache doesn't seem to be
working as expected. For example:

    Q_ASSERT(!cache.contains(buff));
    cache.insert(buff, bytes);
    Q_ASSERT(cache.contains(buff)); // fails!

In the mean time, we'll try copying the data and see whether it's too much of a
performance penalty.

This commit also remove the function from the QGst namespace because that
conflicts with the externally-linked documentation for QtGStreamer. Plus, of
course, it's not actually part of QGst.