Bug squash! Fix intermittent hang at startup
For a while now, I've been having a problem where the application sometimes
hands for up to ~30s at launch, blocking the GUI thread. I finally found out
why! The EegDecoder was calling StripChart::setNStrips() with the value of an
uninitialised variable — the StripChart would go ahead and create up to 2**32
strips (initialising with zero data arrays and everything), causing a hang in
the GUI thread. The correct nStrips value then got set on the next line.

Thanks valgrind!