Change ConfigManager API and update DeviceProxy
The ConfigManager was creating a DeviceDBusAdaptor in its constructor which
in turn calls a virtual method of the ConfigManager. Calling a virtual method
of a class within that class' ctor is not allowed. I've had to extract a method
to expose the ConfigManager to DBus some time after the constructor has been
called — this should be done at the end of OfflineDataSource subclass ctors.

Updated the DeviceProxy to connect to the local session bus if the device
address == "localhost". Thus, when using an OfflineDataSource you must call
DeviceProxy::connectTo("localhost"). This seems a little bit messy, but it'll do
for now.

Also implemented a test to read settings from a mock ConfigManager over DBus.