Stop server when connected in case of prior crash
The server has no way of knowing if the client unexpectedly exits, so will keep
running if the client crashes. Telling the server to stop when the client first
connects helps a little, though it would be nice if the server could detect
client disconnection. Maybe use a watchdog/keepalive?