Refactor Pipeline::setStartTime()
Avoid connect/disconnect complexity. Just leave the watcher connected and have a
flag to check whether the start time has been set yet or not. This introduces a
little runtime overhead since it will call the watcher (that simply returns) on
every sample. The repeated dis/connect was causing issues, though, so I'll live
with the slight performance hit.