Begin refactoring PluginManager
It will be structured around several PluginHosts, each supporting plugins
written in a different language. A PluginHost knows how to list the classes
provided by a plugin and how to instantiate them. The PluginManager delegates
plugin searching to the PluginHosts, maintains the list of found classes, and
provides methods to select an ElementSet from the available classes.

So far the NativeHost is implemented and tested. The PluginManager is only a
stub at the moment.