Elements don't have QObject parents.
Pipeline elements must not have QObject parents, because elements are passed
around by shared pointers. Shared pointers are responsible for deleting their
referents, and QObjects are responsible for deleting their children. If an
object managed by a shared pointer has a QObject parent, there is the
possibility of a double-delete occurring. Remove the optional parent argument to
the pipeline element constructors to prevent this from ever happening.