Improve ElementSet ownership semantics
An ElementSet contains unique_ptrs to elements because it owns them. It exposes
raw pointers as views of these elements. An ElementSet itself should probably
also be referred to by a unique_ptr, but that's not possible if it is to be used
as a parameter of a signal; a shared_ptr is used instead.