Use a class decorator for python element classes
When writing plugins in python, rather than defining a list of exported classes
you now decorate each class to export with `@elapse.export`. This is a bit more
pythonic. It would be nicer to have a metaclass that automatically exports the
classes when they are defined, but that's a bit more effort than it's worth and
may interfere with boost.python's internals.